Выпадающий список открывается по клику на весь блок, а не только по клику на стрелку
